requirements-lock.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
django-filter==2.2.0
coreschema==0.0.4
drf-yasg==1.17.1
psycopg2==2.8.4
django-mptt==0.11.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==1.0.1
django-mptt-urls==2.0.4
djangorestframework==3.10.0
djangorestframework-recursive==0.1.2
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.1
ppretty==1.3
psycopg2==2.8.3
pyparsing==2.4.6
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
